handle errors in relaunch script

+++ relaunchOktaProcess.sh
@@ -1,13 +1,29 @@
 #!/bin/bash
 
+PATH=$PATH:/usr/local/bin
+
 # kill the Quicktime Broadcaster app (find its PID first)
-kill -s SIGTERM $(ps aux | grep Broadcaster | grep -v grep | awk '{print $2}')
+QT=$(ps aux | grep Broadcaster | grep -v grep | awk '{print $2}')
+if [ -n "$QT" ]; then
+	echo "killing QTBroadcaster"
+	kill -s SIGTERM $QT
+else
+	echo "QTBroadcaster not running so not killing"
+fi
 
 # kill the okta camera app (find its PID first)
-kill -s SIGTERM $(ps aux | grep okta_2015 | grep -v grep | awk '{print $2}')
+OVID=$(ps aux | grep okta_2015 | grep -v grep | awk '{print $2}')
+if [ -n "$OVID" ]; then
+	echo "killing okta_2015"
+	kill -s SIGTERM $OVID
+else
+	echo "okta_2015 not running so not killing anything"
+fi
 
 # kill the lunch process but we know its PID
-cat /var/tmp/lunch/master-okta-macoslunch.pid | xargs kill -s SIGTERM
+if [ -f /var/tmp/lunch/master-okta-macoslunch.pid ]; then
+	cat /var/tmp/lunch/master-okta-macoslunch.pid | xargs kill -s SIGTERM
+fi
 
 sleep 15
-./okta-macos.lunch &
+/Users/okta/src/okta/okta-macos.lunch &
